moving functions around
